requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose[cyptography]
passlib[bcrypt]
python-multipart